requirements.txt
# Core dependencies
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.0
pytest-bdd>=6.1.0
pytest-html>=4.0.0
pytest-xdist>=3.5.0
allure-pytest>=2.13.0

# Data handling
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0

# API testing
requests>=2.31.0

# Database
psycopg2-binary>=2.9.0

# Logging and utilities
python-dotenv>=1.0.0
colorlog>=6.7.0
